+++ src/containers/WebComponentLoader.jsx
@@ -14,6 +14,12 @@ import { useProjectPersistence } from "../hooks/useProjectPersistence";
 import { removeUser, setUser } from "../redux/WebComponentAuthSlice";
 import { SettingsContext } from "../utils/settings";
 import { useCookies } from "react-cookie";
+import NewFileModal from "../components/Modals/NewFileModal";
+import ErrorModal from "../components/Modals/ErrorModal";
+import RenameFileModal from "../components/Modals/RenameFileModal";
+import NotFoundModal from "../components/Modals/NotFoundModal";
+import AccessDeniedNoAuthModal from "../components/Modals/AccessDeniedNoAuthModal";
+import AccessDeniedWithAuthModal from "../components/Modals/AccessDeniedWithAuthModal";
 
 const WebComponentLoader = (props) => {
   const loading = useSelector((state) => state.editor.loading);
@@ -41,6 +47,25 @@ const WebComponentLoader = (props) => {
     (state) => state.editor.hasShownSavePrompt,
   );
   const saveTriggered = useSelector((state) => state.editor.saveTriggered);
+  const modals = useSelector((state) => state.editor.modals);
+  const errorModalShowing = useSelector(
+    (state) => state.editor.errorModalShowing,
+  );
+  const newFileModalShowing = useSelector(
+    (state) => state.editor.newFileModalShowing,
+  );
+  const renameFileModalShowing = useSelector(
+    (state) => state.editor.renameFileModalShowing,
+  );
+  const notFoundModalShowing = useSelector(
+    (state) => state.editor.notFoundModalShowing,
+  );
+  const accessDeniedNoAuthModalShowing = useSelector(
+    (state) => state.editor.accessDeniedNoAuthModalShowing,
+  );
+  const accessDeniedWithAuthModalShowing = useSelector(
+    (state) => state.editor.accessDeniedWithAuthModalShowing,
+  );
 
   const [cookies, setCookie] = useCookies(["theme", "fontSize"]);
   const themeDefault = window.matchMedia("(prefers-color-scheme:dark)").matches
@@ -112,6 +137,12 @@ const WebComponentLoader = (props) => {
           sidebarOptions={sidebarOptions}
           hostStyles={hostStyles}
         />
+        {errorModalShowing && <ErrorModal />}
+        {newFileModalShowing && <NewFileModal />}
+        {renameFileModalShowing && modals.renameFile && <RenameFileModal />}
+        {notFoundModalShowing && <NotFoundModal />}
+        {accessDeniedNoAuthModalShowing && <AccessDeniedNoAuthModal />}
+        {accessDeniedWithAuthModalShowing && <AccessDeniedWithAuthModal />}
       </SettingsContext.Provider>
     </>
   ) : (